# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Webhook Delivery Platform API
  version: 0.5.0
  license:
    name: MIT
    identifier: MIT
  description: |
    Durable webhook endpoint, event, and delivery operations.
    持久化 Webhook Endpoint、事件与投递任务 API。
servers:
  - url: /
security:
  - cookieAuth: []
paths:
...
  /api/auth/login:
    post:
      operationId: loginOperator
      summary: Start an operator session / 创建操作者会话
      security: []
      parameters:
        - $ref: "#/components/parameters/CsrfHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Credentials accepted and the session identifier rotated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperatorSession"
        "401":
          $ref: "#/components/responses/InvalidCredentials"
        "403":
          $ref: "#/components/responses/AccessDenied"
  /api/auth/session:
    get:
      operationId: getOperatorSession
      summary: Inspect the current operator session / 查询当前操作者会话
      responses:
        "200":
          description: Authenticated operator identity.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperatorSession"
        "401":
          $ref: "#/components/responses/Unauthenticated"
  /api/auth/logout:
    post:
      operationId: logoutOperator
      summary: End the operator session / 结束操作者会话
      parameters:
        - $ref: "#/components/parameters/CsrfHeader"
      responses:
        "204":
          description: Session invalidated and authentication cookies cleared.
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/AccessDenied"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID
  parameters:
    CsrfHeader:
      name: X-XSRF-TOKEN
      in: header
      required: true
      description: Session-bound token obtained from `GET /api/auth/csrf`.
      schema:
        type: string
...
  responses:
    Unauthenticated:
      description: No valid operator session. Error code is `unauthenticated`.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    InvalidCredentials:
      description: The supplied credentials are invalid. Error code is `invalid_credentials`.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    AccessDenied:
      description: CSRF validation or authorization failed. Error code is `access_denied`.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
...
  schemas:
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 120
        password:
          type: string
          minLength: 1
          maxLength: 512
          writeOnly: true
    OperatorSession:
      type: object
      required: [username]
      properties:
        username:
          type: string
...
    ApiError:
      type: object
      required: [timestamp, status, code, message, fields]
      properties:
        timestamp: { type: string, format: date-time }
        status: { type: integer }
        code: { type: string }
        message: { type: string }
        fields:
          type: object
          additionalProperties: { type: string }
        requestId: { type: [string, "null"] }